LayoutTests/accessibility/gtk/name-from-label-expected.txt

Full Name:
This tests that accessibility objects derive their names from associated labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entry.role is 'AXRole: AXTextField'
PASS entry.title is 'AXTitle: Full Name:'
PASS successfullyParsed is true

TEST COMPLETE
